refactor(pipeline/scrape): extract FieldHunt walkers
Phase 5 commit 4/9 - move the two LIFO stack-based crawlers out
of ScrapeAutoMapper.ts into the dedicated FieldHunt/ cluster:

- FieldHunt/LifoCrawl.ts (106 eff): MAX_ARRAY_DEPTH,
  MAX_DRAIN_ITERATIONS, IStackEntry, IArrayNodeCtx,
  pushArrayChildren, pushObjectChildren, handleArrayNode,
  processStackEntry, processOneLifo, drainLifoStack,
  findFirstArray. Powers ContainerPicker's first-array fallback.

- FieldHunt/TxnHunt.ts (60 eff): MAX_HUNT_DEPTH, IHuntEntry,
  IHuntArrayArgs, processHuntArray, processHuntObject,
  processHuntEntry, drainHuntStack, huntTransactions. Powers
  ContainerPicker's primary multi-array extractor.

Both files use `export default` for their single public symbol
(matches the precedent set by Network/FrozenReplay.ts and others
in the cluster, satisfies import-x/prefer-default-export).

Sub-split into two files so each stays under the per-cluster
max-lines:150 eff cap — a combined FieldHunt.ts measured 166 eff.

ScrapeAutoMapper.ts (462 eff, was 619) drops MIN_TXN_SCORE and
scoreTxnSignature imports since they are now consumed inside the
FieldHunt sub-modules.

Master plan: pipeline-decoupling-master-2026-05-28 / phase-5
Cap: max-lines 150 eff + max-lines-per-function 20 eff
File eff LoC: LifoCrawl.ts=106, TxnHunt.ts=60 (cap 150)
