+++ src/Scrapers/Pipeline/Mediator/Scrape/ScrapeAutoMapper.ts
@@ -32,8 +32,10 @@ import {
   isSearchableObject,
   matchField,
 } from './BfsFieldSearch/BfsFieldSearch.js';
-import { castSearchable, MIN_TXN_SCORE, scoreTxnSignature } from './BfsFieldSearch/TxnSignature.js';
+import { castSearchable } from './BfsFieldSearch/TxnSignature.js';
 import { coerceNumber, coerceString, parseAutoDate } from './Coercion/Coercion.js';
+import findFirstArray from './FieldHunt/LifoCrawl.js';
+import huntTransactions from './FieldHunt/TxnHunt.js';
 
 export type { IMonthChunk } from '../Scrape/ScrapeReplayAction.js';
 export {
@@ -71,179 +73,6 @@ function coerceIdentifier(val: ScalarFieldHit): string | number | false {
 }
 
 // KNOWN_DATE_FORMATS imported from Registry/WK/ScrapeWK.ts
-
-/** Max depth for stack-based array search. */
-const MAX_ARRAY_DEPTH = 15;
-
-/** Stack entry for LIFO array search. */
-interface IStackEntry {
-  readonly node: unknown;
-  readonly depth: number;
-}
-
-/** Context for handleArrayNode. */
-interface IArrayNodeCtx {
-  readonly collected: unknown[];
-  readonly stack: IStackEntry[];
-  readonly node: unknown[];
-  readonly depth: number;
-}
-
-/**
- * Push object items from an array onto the stack.
- * @param stack - Exploration stack.
- * @param items - Array items to push.
- * @param depth - Current depth.
- * @returns Number of items pushed.
- */
-function pushArrayChildren(
-  stack: IStackEntry[],
-  items: readonly UntypedValue[],
-  depth: number,
-): number {
-  const objects = items.filter((item): boolean => typeof item === 'object' && item !== null);
-  for (const obj of objects) {
-    stack.push({ node: obj, depth: depth + 1 });
-  }
-  return objects.length;
-}
-
-/**
- * Push child values of an object onto the stack.
- * @param stack - Exploration stack.
- * @param obj - Object whose values to push.
- * @param depth - Current depth.
- * @returns Number of items pushed.
- */
-function pushObjectChildren(
-  stack: IStackEntry[],
-  obj: Record<string, unknown>,
-  depth: number,
-): number {
-  const values = Object.values(obj);
-  for (const value of values) {
-    stack.push({ node: value, depth: depth + 1 });
-  }
-  return values.length;
-}
-
-/**
- * Handle an array node during LIFO traversal.
- * @param ctx - Array node context.
- * @returns True if items were collected.
- */
-function handleArrayNode(ctx: IArrayNodeCtx): boolean {
-  if (ctx.node.length === 0) return false;
-  const firstNode = ctx.node[0];
-  const score = scoreTxnSignature(firstNode);
-  if (score < MIN_TXN_SCORE) {
-    pushArrayChildren(ctx.stack, ctx.node, ctx.depth);
-    return false;
-  }
-  for (const item of ctx.node) ctx.collected.push(item);
-  return true;
-}
-
-/**
- * Process one stack entry: dispatch array vs object.
- * @param entry - Stack entry to process.
- * @param collected - Accumulator for items.
- * @param stack - Exploration stack.
- * @returns True if the entry was processed.
- */
-function processStackEntry(
-  entry: IStackEntry,
-  collected: UntypedValue[],
-  stack: IStackEntry[],
-): boolean {
-  if (entry.depth > MAX_ARRAY_DEPTH) return false;
-  if (Array.isArray(entry.node)) {
-    return handleArrayNode({
-      collected,
-      stack,
-      node: entry.node,
-      depth: entry.depth,
-    });
-  }
-  const isObj = typeof entry.node === 'object' && entry.node !== null;
-  if (!isObj) return false;
-  const record = entry.node as Record<string, unknown>;
-  pushObjectChildren(stack, record, entry.depth);
-  return true;
-}
-
-/**
- * Find the best transaction array via LIFO crawl.
- * Falls back to BFS if no high-score arrays found.
- * @param obj - Parsed API response.
- * @returns Flattened array of transaction-like items.
- */
-/**
- * Process one LIFO iteration — pop last entry and process it.
- * @param stack - Mutable stack.
- * @param collected - Mutable accumulator.
- * @returns True if stack is now empty.
- */
-function processOneLifo(stack: IStackEntry[], collected: UntypedValue[]): boolean {
-  // `Array.pop` returns `undefined` on an empty stack, consolidating the
-  // "stack is drained" signal into a single check.
-  const entry = stack.pop();
-  if (entry === undefined) return true;
-  processStackEntry(entry, collected, stack);
-  return stack.length === 0;
-}
-
-/** Hard cap on iterations — defends against pathological cycles. */
-const MAX_DRAIN_ITERATIONS = 1_000_000;
-
-/**
- * LIFO drain — iterative loop via `Array.every` short-circuit.
- * Processes every stack entry until `processOneLifo` reports the
- * stack is empty. Iterative (was tail-recursive) so adversarial
- * bodies with tens of thousands of nested items (e.g. Lottie
- * animation JSON) cannot blow Node's stack budget.
- * @param stack - Mutable stack.
- * @param collected - Mutable accumulator.
- * @returns Collected items.
- */
-function drainLifoStack(stack: IStackEntry[], collected: UntypedValue[]): readonly UntypedValue[] {
-  // `every` short-circuits when the predicate returns false. Each
-  // step calls `processOneLifo`; when it returns true (stack drained)
-  // the negation short-circuits `every` and the loop ends.
-  const iterations: readonly number[] = Array.from(
-    { length: MAX_DRAIN_ITERATIONS },
-    (_unused, i): number => i,
-  );
-  iterations.every((): boolean => !processOneLifo(stack, collected));
-  return collected;
-}
-
-/**
- * Find the first array of objects in a nested structure via LIFO traversal.
- * @param obj - Root API record to search.
- * @returns First array of searchable items found.
- */
-function findFirstArray(obj: ApiRecord): readonly UntypedValue[] {
-  const initial: IStackEntry = { node: obj, depth: 0 };
-  const stack: IStackEntry[] = [initial];
-  const collected: UntypedValue[] = [];
-  drainLifoStack(stack, collected);
-  if (collected.length > 0) {
-    LOG.debug({ message: `findFirstArray: collected ${String(collected.length)} items` });
-    return collected;
-  }
-  LOG.debug({
-    message: 'findFirstArray: falling back to BFS',
-  });
-  const allObjects = flattenObjectTree(obj);
-  const arrays = allObjects.map((o): readonly UntypedValue[] | false => {
-    const arr = Object.values(o).find(Array.isArray);
-    if (arr) return arr as readonly UntypedValue[];
-    return false;
-  });
-  const hit = arrays.find((a): a is readonly UntypedValue[] => a !== false);
-  return hit ?? [];
-}
 
 /** Shekel currency aliases from WK. */
 const SHEKEL_ALIASES = new Set(WK.shekelAliases);
@@ -794,105 +623,6 @@ function extractValidIdentifier(record: ApiRecord): string {
 function extractAccountIds(responseBody: ApiRecord): readonly string[] {
   const records = extractAccountRecords(responseBody);
   return records.map(extractValidIdentifier).filter(Boolean);
-}
-
-/** Max depth for transaction hunting. */
-const MAX_HUNT_DEPTH = 20;
-
-/** Stack entry for iterative tree walk. */
-interface IHuntEntry {
-  readonly val: unknown;
-  readonly depth: number;
-}
-
-/** Bundled args for array processing. */
-interface IHuntArrayArgs {
-  readonly objects: readonly unknown[];
-  readonly depth: number;
-  readonly collected: ApiRecord[];
-  readonly stack: IHuntEntry[];
-}
-
-/**
- * Process one array node — collect if txn-like, else push children.
- * @param args - Bundled hunt array arguments.
- * @returns True if collected.
- */
-function processHuntArray(args: IHuntArrayArgs): boolean {
-  const { objects, depth, collected, stack } = args;
-  if (objects.length === 0) return false;
-  const firstObj = objects[0] as ApiRecord;
-  const score = scoreTxnSignature(firstObj);
-  if (score >= MIN_TXN_SCORE) {
-    collected.push(...(objects as ApiRecord[]));
-    return true;
-  }
-  const children = objects.map((o): IHuntEntry => ({ val: o, depth: depth + 1 }));
-  stack.push(...children);
-  return false;
-}
-
-/**
- * Process one object node — push child values onto stack.
- * @param record - Object to expand.
- * @param depth - Current depth.
- * @param stack - Mutable stack.
- * @returns True after pushing.
- */
-function processHuntObject(
-  record: Record<string, unknown>,
-  depth: number,
-  stack: IHuntEntry[],
-): boolean {
-  const children = Object.values(record)
-    .filter((v): boolean => typeof v === 'object' && v !== null)
-    .map((v): IHuntEntry => ({ val: v, depth: depth + 1 }));
-  stack.push(...children);
-  return true;
-}
-
-/**
- * Process one stack entry — dispatch array vs object.
- * @param entry - Entry to process.
- * @param collected - Mutable collector.
- * @param stack - Mutable stack.
- * @returns True if processed.
- */
-function processHuntEntry(entry: IHuntEntry, collected: ApiRecord[], stack: IHuntEntry[]): boolean {
-  if (entry.depth > MAX_HUNT_DEPTH) return false;
-  const { val, depth } = entry;
-  if (Array.isArray(val)) {
-    const objects = (val as unknown[]).filter((v): boolean => typeof v === 'object' && v !== null);
-    return processHuntArray({ objects, depth, collected, stack });
-  }
-  if (typeof val === 'object' && val !== null) {
-    return processHuntObject(val as Record<string, unknown>, depth, stack);
-  }
-  return false;
-}
-
-/**
- * Drain the hunt stack — process entries until empty.
- * @param stack - Mutable stack.
- * @param collected - Mutable collector.
- * @returns Collected transaction items.
- */
-function drainHuntStack(stack: IHuntEntry[], collected: ApiRecord[]): readonly ApiRecord[] {
-  if (stack.length === 0) return collected;
-  const entry = stack.pop();
-  if (entry) processHuntEntry(entry, collected, stack);
-  return drainHuntStack(stack, collected);
-}
-
-/**
- * Stack-based iterative transaction hunter.
- * Walks the response tree. Collects arrays whose items score as transactions.
- * @param responseBody - Raw API response.
- * @returns Flat array of transaction-like items.
- */
-function huntTransactions(responseBody: ApiRecord): readonly ApiRecord[] {
-  const stack: IHuntEntry[] = [{ val: responseBody, depth: 0 }];
-  return drainHuntStack(stack, []);
 }
 
 /**
